move reusable utiliy functions to utilities folder
