+++ packages/eth-graphql/src/utilities/formatToEntityName.ts
