+++ packages/eth-graphql/src/utilities/formatToSignature.ts
